api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-storage==2.17.0

python-dotenv==1.0.1
